+++ tests/suite/correctness_FO.ml
@@ -88,3 +88,13 @@ let () = declare "typeabbrv10"
   ~expectation:(Output (Str.regexp "3 has type int but is used with type (foo X[0-9]+)"))
   ~description:"type abbreviations and error messages"
   ()
+let () = declare "typeabbrv11"
+  ~source_elpi:"typeabbrv11.elpi"
+  ~expectation:(Output (Str.regexp "x has type string but is used with type int"))
+  ~description:"type abbreviations and error messages"
+  ()
+let () = declare "typeabbrv12"
+  ~source_elpi:"typeabbrv12.elpi"
+  ~expectation:(Output (Str.regexp "x has type string but is used with type y"))
+  ~description:"type abbreviations and error messages"
+  ()